requirements.txt
requests
clickhouse-connect
fastapi
pandas
python-dotenv
uvicorn
pydantic
sqlalchemy
psycopg2